ALL WARNINGS FIXED

+++ Eclipse/src/ECS/SystemManager/Systems/System/CameraSystem.cpp
@@ -24,7 +24,7 @@ void Eclipse::CameraSystem::Update()
 	ZoneScopedN("Camera System")
 
 	engine->Timer.SetName({ SystemName::CAMERA });
-	engine->Timer.tracker.system_start = glfwGetTime();
+	engine->Timer.tracker.system_start = static_cast<float>(glfwGetTime());
 	for (auto& it : mEntities)
 	{
 		auto& _camera = engine->world.GetComponent<CameraComponent>(it);
@@ -46,8 +46,7 @@ void Eclipse::CameraSystem::Update()
 		engine->gCamera.ComputeViewMtx(_camera, _transform);
 		engine->gCamera.ComputePerspectiveMtx(_camera);
 	}
-	engine->Timer.tracker.system_end = glfwGetTime();
-
+	engine->Timer.tracker.system_end = static_cast<float>(glfwGetTime());
 	engine->Timer.UpdateTimeContainer(engine->Timer.tracker);
 
 	FrameMark
