+++ Eclipse/src/ECS/SystemManager/Systems/System/GridSystem/GridSystem.cpp
@@ -58,7 +58,7 @@ namespace Eclipse
             }
         }
 
-        engine->Timer.tracker.system_end = glfwGetTime();
+        engine->Timer.tracker.system_end = static_cast<float>(glfwGetTime());
         engine->Timer.UpdateTimeContainer(engine->Timer.tracker);
 
         FrameMark
