+++ Eclipse/src/Graphics/Grid/AABBTree.cpp
@@ -331,7 +331,7 @@ namespace Eclipse
         return overlaps;
     }
 
-    unsigned int AABBTree::NumberOfIntersections(std::vector<unsigned int>& in)
+    size_t AABBTree::NumberOfIntersections(std::vector<unsigned int>& in)
     {
         return in.size();
     }
